add UserResponse , refactor mapper, update UserInputAdapter
