+++ src/main/java/com/example/demo/inputs/UserInputAdapter.java
@@ -2,6 +2,8 @@ package com.example.demo.inputs;
 
 import com.example.demo.domain.models.User;
 import com.example.demo.domain.ports.UserInputPort;
+import com.example.demo.inputs.models.UserResponse;
+import com.example.demo.outputs.UserMapper;
 import lombok.AllArgsConstructor;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -13,15 +15,13 @@ import java.util.List;
 @AllArgsConstructor
 @RestController
 @RequestMapping(path = "api/v1/users")
-
 public class UserInputAdapter {
     private final UserInputPort userInputPort;
 
-  @GetMapping
-  public List<UserResponse>  getUsers() {
+    @GetMapping
+    public List<UserResponse> getUsers() {
         List<User> users = userInputPort.getUsers();
-        return null;
-  }
-
-
+        List <UserResponse> userResponses = UserMapper.mapUserListToUserReponse(users);
+        return userResponses;
+    }
 }
