+++ src/main/java/com/example/demo/inputs/models/UserResponse.java
@@ -1,0 +1,11 @@
+package com.example.demo.inputs.models;
+
+public record UserResponse(
+        Long id,
+        String username,
+        String firstName,
+        String lastName,
+        String email,
+        String phoneNumber
+) {}
+
