+++ src/main/java/com/example/demo/outputs/repositories/UserRepository.java
@@ -1,4 +1,4 @@
-package com.example.demo.user;
+package com.example.demo.outputs.repositories;
 
 import com.example.demo.outputs.entities.UserEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
